further cleanup of strategy classes
