+++ lumibot/brokers/interactive_brokers_rest.py
@@ -1,17 +1,14 @@
 import logging
 from termcolor import colored
-from lumibot.brokers import Broker
-from lumibot.entities import Order, Asset, Position
-from lumibot.data_sources import InteractiveBrokersRESTData
+from ..brokers import Broker
+from ..entities import Order, Asset, Position
+from ..data_sources import InteractiveBrokersRESTData
 import datetime
 from decimal import Decimal
 from math import gcd
 import re
-import ssl
-import time
-import json
 import traceback
-from lumibot.trading_builtins import PollingStream
+from ..trading_builtins import PollingStream
 
 TYPE_MAP = dict(
     stock="STK",
